Upgrade api-stress-test to production-grade quality
Critical fixes:
- Limit response body drain to 1MB to prevent worker blocking
- Fix ParseData to return error on malformed input instead of silent skip
- Add panic recovery in worker goroutines
- Add input validation for timeout, rate, and concurrency
- Fix signal handler cleanup to prevent goroutine leaks
- Guard against division by zero in req/s calculation

New features:
- Live progress bar with real-time req/s and completion percentage
- Latency histogram (ASCII in text, array in JSON output)
- Transport tuning flags: --insecure/-k, --disable-keepalive, --disable-redirects
- Response validation: --expect-status and --expect-body flags
- Warm-up phase via --warmup flag with signal handling and backoff
- Throughput timeline (per-second req/s) in JSON output
- ANSI colors for text output with auto-detection for terminal/pipe
- Error normalization for better Top Errors grouping

Architecture improvements:
- Extract output formatting to internal/ui/ package
- Reservoir sampling (10K cap) for bounded-memory percentile calculation
- RunStressTest accepts io.Writer for testability

Test coverage:
- Add 8 unit tests for ExecuteRequest (status codes, timeout, headers, body)
- Add 8 integration tests for RunStressTest (success, failure, JSON, duration,
  rate limiting, body, expect-status, expect-body)
- Unify Statistics.Total to int64 for consistency

+++ api-stress-test/internal/stats/collector_test.go
@@ -125,7 +125,7 @@ func TestCollectorConcurrency(t *testing.T) {
 	wg.Wait()
 	stat := c.GetStatistics()
 
-	expected := numGoroutines * recordsPerGoroutine
+	expected := int64(numGoroutines * recordsPerGoroutine)
 	if stat.Total != expected {
 		t.Errorf("total = %d, want %d", stat.Total, expected)
 	}
